requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
supabase>=2.5.0
python-dotenv>=1.0.1